#4: Updated Folder Structure
