+++ lib/core/property_holder.dart
@@ -19,11 +19,9 @@ extension PropertyHolderListExtension on List<PropertyHolder> {
 class PropertyHolder<T> {
   final String id;
   final Widget widget;
-  final Function(T) onChanged;
 
   const PropertyHolder({
     required this.id,
     required this.widget,
-    required this.onChanged,
   });
 }
