--- lib/settings/base.dart
@@ -1,3 +1,0 @@
-abstract class Property<T>{
-  T call();
-}
