--- lib/settings/border_radius.dart
@@ -1,7 +1,0 @@
-import 'package:flutter/material.dart';
-
-class BorderRadiusProperties {
-  double radius = 50;
-
-  BorderRadius get borderRadius => BorderRadius.circular(radius);
-}
