--- lib/settings/offset.dart
@@ -1,8 +1,0 @@
-import 'package:flutter/material.dart';
-
-class OffsetProperties {
-  double x = 10;
-  double y = 10;
-
-  Offset get offset => Offset(x, y);
-}
